+++ Tests/MeterTests/MeterTests.swift
@@ -1,0 +1,14 @@
+//
+//  MeterTests.swift
+//  Rhythm
+//
+//  Created by James Bean on 4/26/17.
+//
+//
+
+import XCTest
+import Meter
+
+class MeterTests: XCTestCase {
+
+}
