+++ Tests/MusicModelTests/MeterTempoTests.swift
@@ -1,15 +1,16 @@
 //
-//  MeterTests.swift
-//  Rhythm
+//  MeterTempoTests.swift
+//  MusicModel
 //
-//  Created by James Bean on 4/26/17.
-//
+//  Created by James Bean on 8/26/17.
 //
 
 import XCTest
 import Rhythm
+import Tempo
+import Meter
 
-class MeterTests: XCTestCase {
+class MeterTempoTests: XCTestCase {
 
     func testTempoInit() {
         _ = Tempo(78, subdivision: 4)
