+++ Tests/TempoTests/EasingEvaluateTests.swift
@@ -6,7 +6,7 @@
 //
 //
 import XCTest
-@testable import Rhythm
+@testable import Tempo
 
 class EasingEvaluateTests: XCTestCase {
