+++ Tests/TempoTests/TempoInterpolationCollectionTests.swift
@@ -8,7 +8,7 @@
 
 import XCTest
 import Math
-@testable import Rhythm
+@testable import Tempo
 
 class StratumTests: XCTestCase {
